Rydder i filstruktur for serveren. Pretty-logging ved bygg
