+++ .gitignore
@@ -48,7 +48,7 @@ yarn-error.log*
 
 /server/.dist/
 /server/tsconfig.tsbuildinfo
-/server/__next-test-dummy/.next/
+/server/__test-utils/next-dummy/.next/
 
 # VS Code
 .vscode
