+++ server/__test-utils/utils.ts
@@ -1,0 +1,8 @@
+import next from 'next';
+import path from 'path';
+
+export const __getNextTestApp = () =>
+    next({
+        conf: {},
+        dir: path.join(__dirname, 'next-dummy'),
+    });
