+++ server/src/req-handlers/invalidate-all.ts
@@ -1,11 +1,11 @@
 import { RequestHandler } from 'express';
-import CustomCacheHandler from 'cache/custom-cache-handler';
+import PageCacheHandler from 'cache/page-cache-handler';
 import { logger } from 'srcCommon/logger';
 
 export const handleInvalidateAllReq: RequestHandler = async (req, res) => {
     const { eventid } = req.headers;
 
-    await new CustomCacheHandler().clear();
+    await new PageCacheHandler().clear();
 
     const msg = `Cleared page cache - event id ${eventid}`;
     logger.info(msg);
