+++ server/src/server-setup/server-setup-failover.ts
