+++ server/src/server-setup/server-setup.ts
@@ -7,9 +7,9 @@ import { handleInvalidatePathsReq } from 'req-handlers/invalidate-paths';
 import { setCacheKey } from 'req-handlers/set-cache-key';
 import { handleInvalidateAllReq } from 'req-handlers/invalidate-all';
 import { handleGetPendingResponses } from 'req-handlers/pending-responses';
-import { serverSetupDev } from 'server-setup-dev';
+import { serverSetupDev } from 'server-setup/server-setup-dev';
 import { logger } from 'srcCommon/logger';
-import { redisCache } from 'cache/custom-cache-handler';
+import { redisCache } from 'cache/page-cache-handler';
 
 // Set the no-cache header on json files from the incremental cache to ensure
 // data requested during client side navigation is always validated if cached
